Add more granular unit tests
