+++ todo_app/todo_item.py
@@ -1,4 +1,4 @@
-class Item:
+class TodoItem:
     def __init__(self, id, name, status='To Do'):
         self.id = id
         self.name = name
